requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
requests==2.31.0
pillow==10.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
stripe==7.8.0
razorpay==1.3.0
twilio==8.10.0
redis==5.0.1
celery==5.3.4
posthog==3.0.2
pydantic==2.5.0
pydantic-settings
httpx==0.25.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
boto3==1.34.0
psycopg2-binary
openai
